refactor: 입출력 함수 추출
