+++ src/utils/Console.js
@@ -1,0 +1,28 @@
+import readline from 'readline';
+
+class Console {
+  static readLineAsync(query) {
+    return new Promise((resolve, reject) => {
+      if (arguments.length !== 1) {
+        reject(new Error('arguments must be 1'));
+      }
+
+      if (typeof query !== 'string') {
+        reject(new Error('query must be string'));
+      }
+
+      const rl = readline.createInterface({
+        input: process.stdin,
+        output: process.stdout,
+      });
+
+      rl.question(query, (input) => {
+        rl.close();
+        console.log(input);
+        resolve(input);
+      });
+    });
+  }
+}
+
+export default Console;
