+++ src/view/Input.js
@@ -1,31 +1,8 @@
-import readline from 'readline';
+import Console from '../utils/Console.js';
 
 class Input {
-  readLineAsync(query) {
-    return new Promise((resolve, reject) => {
-      if (arguments.length !== 1) {
-        reject(new Error('arguments must be 1'));
-      }
-
-      if (typeof query !== 'string') {
-        reject(new Error('query must be string'));
-      }
-
-      const rl = readline.createInterface({
-        input: process.stdin,
-        output: process.stdout,
-      });
-
-      rl.question(query, (input) => {
-        rl.close();
-        console.log(input);
-        resolve(input);
-      });
-    });
-  }
-
   async inputCarName() {
-    const input = await this.readLineAsync(
+    const input = await Console.readLineAsync(
       '경주할 자동차 이름을 입력하세요(이름은 쉼표(,)를 기준으로 구분).'
     );
 
@@ -33,7 +10,7 @@ class Input {
   }
 
   async inputTryCount() {
-    const input = await this.readLineAsync('시도할 횟수는 몇 회인가요?');
+    const input = await Console.readLineAsync('시도할 횟수는 몇 회인가요?');
 
     return input;
   }
